container: add a docker to podman playbook migration
This playbook is aims to ease the rhel7 to rhel8 upgrade.

Since rhel8 will be podman only, the idea is to prepare all systemd
units to use podman before the system is upgraded.
Once the OS is upgraded to rhel8, it will be able to start ceph daemons
via podman.

Calling the `infrastructure-playbooks/docker2podman.yml` playbook with
`--tags docker2podman` will reuse all existing roles in order to render
systemd units. However it could be very long to run the whole playbook
just for few tasks, the idea here is to play with tags so we play only
required tasks.

The `docker2podman` variable is set to `True` so handlers won't be
trigger to avoid a restart of all services.

Signed-off-by: Guillaume Abrioux <gabrioux@redhat.com>